refactor: change uint64_t in sprite to size_t

+++ core/src/graphics/sprite.h
@@ -9,8 +9,8 @@ namespace graphics {
 struct Sprite {
   std::vector<uint8_t> buffer;
 
-  uint64_t height, width;
-  uint64_t x, y;
+  size_t height, width;
+  size_t x, y;
 };
 
 }  // namespace graphics
